Added main section and modified layout.
Added main section to html. This section contains the intro and the form.

Changed several css properties. Added css selector for main section.
Main is designed as a flex box, column with all the elements centered along main and cross axes.
Changed the width and height of the body element, to make the page larger.